+++ Navigation/ProfileViewController.swift
@@ -8,6 +8,12 @@
 import UIKit
 
 class ProfileViewController: UIViewController {
+    private lazy var profileHeader: ProfileHeaderView = {
+        let profileHeader = ProfileHeaderView(frame:CGRect(x: 0, y: 88, width: 100, height: 100))
+        profileHeader.translatesAutoresizingMaskIntoConstraints = false
+        return profileHeader
+    }()
+    
     
     private lazy var tableView: UITableView = {
         let tableView = UITableView(frame: .zero, style: .plain)
@@ -18,6 +24,7 @@ class ProfileViewController: UIViewController {
         tableView.register(UITableViewCell.self, forCellReuseIdentifier: "DefaultCell")
         tableView.dataSource = self
         tableView.register(PostTableViewCell.self, forCellReuseIdentifier: "PostCell")
+       
         
      return tableView
     }()
@@ -25,31 +32,46 @@ class ProfileViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         self.setupView()
-    
+        self.setupGesture()
+        
     }
 
    private var posts:[Post] = [post1, post2, post3]
     
     private func setupView(){
         self.view.backgroundColor = .systemBackground
+        self.view.addSubview(profileHeader)
         self.view.addSubview(tableView)
         
         NSLayoutConstraint.activate([
-            self.tableView.topAnchor.constraint(equalTo: self.view.topAnchor),
+            
+            self.profileHeader.topAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.topAnchor),
+            self.profileHeader.leadingAnchor.constraint(equalTo: self.view.leadingAnchor),
+            self.profileHeader.trailingAnchor.constraint(equalTo: self.view.trailingAnchor),
+            self.profileHeader.heightAnchor.constraint(equalToConstant: 220),
+            
+            self.tableView.topAnchor.constraint(equalTo: self.profileHeader.bottomAnchor),
             self.tableView.trailingAnchor.constraint(equalTo: self.view.trailingAnchor),
             self.tableView.leadingAnchor.constraint(equalTo: self.view.leadingAnchor),
-            self.tableView.bottomAnchor.constraint(equalTo: self.view.bottomAnchor)
-        
-        
-        
-        
+            self.tableView.bottomAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.bottomAnchor)
+            
         ])
     }
     
+    private func setupGesture() {
+            let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboard))
+            self.view.addGestureRecognizer(tapGesture)
+        }
+        @objc private func hideKeyboard() {
+            self.view.endEditing(true)
+    
+        }
    
 }
 
-extension ProfileViewController: UITableViewDataSource {
+extension ProfileViewController: UITableViewDataSource, UITableViewDelegate {
+    
+    
    
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
